Added `prior_index` plot by default, fixed plot limit bug
I had assumed that setting the plot limits to `(None, None)` would
keep autoscaling for that axis. However that is *not* the case; it
turns of autoscaling and fixes the current plot limits. This caused
a problem for `prior_index`, because the limits got fixed to `(0, 1)`
which basically covers just the first two index values.

To correct this, the `Limits.get_limit` method returns a third value,
which is a boolean indicating whether or not axis autoscaling should
be kept on.