Bugfix for edge case where context file has no data outside the main file
We had a case where a site had previously uploaded just data for 2018 & 2021,
then later uploaded data from ~2012 to 2016 and 2022. This meant that the
concatenated file didn't actually have any data outside the times covered
by the main file. When the get data functions returned an empty array, a `np.min`
operation errored. Now we check that both the context and reference data
(if requested) have at least one data point which they are allowed, and if not,
do not include them.